petsc/init: Work around segfaults in MPI_Init
OpenMPI might be deciding that one of arguments we're passing to PETSc
is meaningful to it, so meaningful that it decides to crash.  I am too
lazy at the moment to understand which argument this is, since it's
cleaner to not pass a bunch of weird arguments to MPI_Init anyways.